update: updated database primary keys
The method of creating a pair of keys (number as private key for database and uuid as public key)
in commit d4d68af7cdf00372320f2ae4d3aa19c304bb6d15
for each entity in database is not the best one.
It increases complexity when we need to connect two or more tables, because we must use public keys.

In addition as `Question` and `Answer` entities could exist without owner id (because that's how forums or sites like StackOverflow works)
when user deletes his account, database can make related private owner keys as `null` for question and answer.
Hovewer we also need to make public related owner ids as null in question and answer (but we can't make it as with primary keys).
So it was decided to figure out which entity requires which kind of id.

Id's could be two kinds:
- sequential (like id:1, id:2, id:3, etc)
- unique (like uuid)

Sequential ids require less memory, they are simple and predictable which makes database operations like select, insert update much faster.
Hovewer because of they are more predictable, it could not be safe when we don't want other users read information they shouldn't.
Also it could be more complicated to organize sequential ids in distributed systems.

Unique ids which are generated with UUIDv4 or other methods are unique, which makes them almost unpredictable.
They could be used in distributed systems because they are unique.
Hovewer they require much more memory. Database operations and indexing could be slower because they are unpredictable.

More about UUID and sequential ids in these articles:
- https://www.baeldung.com/uuid-vs-sequential-id-as-primary-key
- https://planetscale.com/blog/the-problem-with-using-a-uuid-primary-key-in-mysql

So for the project it was decided to make these changes of entity ids:
- Left `UserCreds` and `User` entities with UUID
- Make `Question`, `Answer`, `Tag` ids sequential
- Make id type for UUID as `ByteA` because standard UUID requires 16 bytes of memory to store (and we don't need to store it as string)
- Make sequential id type as integer with auto increment

As we need to combine using UUID as string and sequential id as number,
I decided to translate them to string type because there could be different methods of generating ids and
to make business logic less reliable to type of id in database.

- With changes of entity and database tables ids we almost have not changed business logic (except for other changes).
- Repository implementations were changed according to new id types
- Application side and dtos were also changed

1. As was mentioned `Question` and `Answer` entities aren't required to have owner.
So foreign keys were changed in database and on owner deleting keys in `Question` and `Answer` tables will be null.
Also dtos, mappers and adapter were changed in application and domain sides

2. Renamed `config` folder with `envs` to `environment` because there are already `config` folder in `src`

3. Created number validator to validate number ids

4. As we can't generate sequential ids on server side I added `returnId` option
when creating `Question`, `QuestionUser`, `Answer`, `AnswerUser` and `Tag` entities.